RCD write progress - NPE related to ByteOp. Also need to consider merging TestByteOperation and WriteByteOperation

+++ src/main/java/lmr/randomizer/rcd/GameObject.java
@@ -9,8 +9,8 @@ import java.util.List;
 public class GameObject {
     private short id;
 
-    private short x;
-    private short y;
+    private int x;
+    private int y;
 
     private List<TestByteOperation> testByteOperations = new ArrayList<>();
     private List<WriteByteOperation> writeByteOperations = new ArrayList<>();
@@ -24,19 +24,19 @@ public class GameObject {
         this.id = id;
     }
 
-    public short getX() {
+    public int getX() {
         return x;
     }
 
-    public void setX(short x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public short getY() {
+    public int getY() {
         return y;
     }
 
-    public void setY(short y) {
+    public void setY(int y) {
         this.y = y;
     }
